Make tx order type from u128 to u64 (#1148)
* make tx order type from u128 to u64

* revert AuthenticationKey serde_as modify

* fixup name mapping

+++ crates/rooch-rpc-server/src/server/rooch_server.rs
@@ -224,7 +224,7 @@ impl RoochAPIServer for RoochServer {
 
     async fn get_transactions_by_order(
         &self,
-        cursor: Option<StrView<u128>>,
+        cursor: Option<StrView<u64>>,
         limit: Option<StrView<u64>>,
     ) -> RpcResult<TransactionWithInfoPageView> {
         let last_sequencer_order = self
@@ -236,7 +236,7 @@ impl RoochAPIServer for RoochServer {
         let limit_of = limit.map(Into::into).unwrap_or(DEFAULT_RESULT_LIMIT);
         let cursor = cursor.map(|v| v.0);
         let start = cursor.unwrap_or(0);
-        let end = min(start + ((limit_of + 1) as u128), last_sequencer_order + 1);
+        let end = min(start + (limit_of + 1), last_sequencer_order + 1);
 
         let mut tx_orders: Vec<_> = if cursor.is_some() {
             ((start + 1)..=end).collect()
